# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_acerowl_lockin_backend_internal_models.LessonActivity:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      title:
        type: string
      topic_name:
        type: string
    type: object
...
  github_com_acerowl_lockin_backend_internal_models.Option:
    properties:
      explanation:
        type: string
      id:
        type: string
      index:
        type: integer
      is_correct:
        type: boolean
      label:
        type: string
      questionId:
        type: string
    type: object
  github_com_acerowl_lockin_backend_internal_models.Question:
    properties:
      answer:
        type: string
      concept_tags:
        items:
          type: string
        type: array
      createdAt:
        type: string
      explanation:
        type: string
      id:
        type: string
      index:
        type: integer
      lessonId:
        type: string
      nodeId:
        type: string
      options:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Option'
        type: array
      question:
        type: string
      type:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.QuestionType'
    type: object
  github_com_acerowl_lockin_backend_internal_models.QuestionType:
    enum:
    - mcq
    - true_false
    - fill_blank
    - short_answer
    - speech
    type: string
    x-enum-varnames:
    - MCQ
    - TrueFalse
    - FillBlank
    - ShortAnswer
    - Speech
  github_com_acerowl_lockin_backend_internal_models.QuizActivity:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      topic_name:
        type: string
    type: object
...
  github_com_acerowl_lockin_backend_internal_models.UserActivityData:
    properties:
      day:
        type: string
      lessons:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.LessonActivity'
        type: array
      quizes:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.QuizActivity'
        type: array
      total_time:
        description: in seconds
        type: integer
    type: object
...
  internal_handlers.CompleteSessionData:
    properties:
      status:
        example: completed
        type: string
    type: object
  internal_handlers.CompleteSessionReq:
    properties:
      answers:
        items:
          $ref: '#/definitions/internal_handlers.UserAnswer'
        type: array
      topic_id:
        type: string
    type: object
  internal_handlers.CompleteSessionResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.CompleteSessionData'
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  internal_handlers.GetUserActivityResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.UserActivityInfo'
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.RateReviewCardReq:
    properties:
      quality:
        description: 0-5 SM-2 quality score
        example: 3
        type: integer
    type: object
...
  internal_handlers.SocraticFollowUpReq:
    properties:
      answer:
        example: Because momentum carries the optimizer past shallow local minima
        type: string
      question_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  internal_handlers.StartSessionData:
    properties:
      questions:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Question'
        type: array
      session_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  internal_handlers.StartSessionReq:
    properties:
      interleave:
        description: 'mixed-review mode: mixes types and injects due review cards
          from other topics'
        example: false
        type: boolean
      lesson_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      quiz_mode:
        description: options, text, lesson
        example: options
        type: string
      topic_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  internal_handlers.StartSessionResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.StartSessionData'
      success:
        example: true
        type: boolean
    type: object
  internal_handlers.SuccessResponse:
    properties:
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.UserActivityInfo:
    properties:
      active_streak:
        example: 3
        type: integer
      activity:
        items:
          $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.UserActivityData'
        type: array
      highest_streak:
        example: 10
        type: integer
    type: object
  internal_handlers.UserAnswer:
    properties:
      answer:
        type: string
      confidence:
        description: '"low" | "med" | "high"'
        type: string
      question:
        $ref: '#/definitions/github_com_acerowl_lockin_backend_internal_models.Question'
    type: object
...
host: localhost:8080
info:
  contact:
    name: LockIn Support
    url: https://github.com/acerowl/lockin
  description: Adaptive learning backend — quiz generation, sessions, and progress
    tracking.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: LockIn API
  version: "1.0"
paths:
...
  /reviews/{id}/rate:
    post:
      consumes:
      - application/json
      description: Applies SM-2 scheduling based on a quality score (0-5). Higher
        is better recall.
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      - description: Quality score 0-5
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RateReviewCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: Updated card with next schedule
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rate a review card
      tags:
      - reviews
  /reviews/due:
    get:
      description: Returns the user's cards that are due for review, optionally filtered
        by topic.
      parameters:
      - description: Filter by topic (optional)
        in: query
        name: topic_id
        type: string
      - description: Max number of cards to return (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Due review cards
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get due review cards
      tags:
      - reviews
  /reviews/due/count:
    get:
      description: Returns the total count of review cards due right now for the user.
      produces:
      - application/json
      responses:
        "200":
          description: Due count
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get number of due review cards
      tags:
      - reviews
  /reviews/retention:
    get:
      description: Returns retention data for each topic over the last N days (default
        7).
      parameters:
      - description: Number of days to look back (default 7)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retention series by topic
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get retention by topic over time
      tags:
      - reviews
  /reviews/stats:
    get:
      description: Returns retention-by-interval, weak concepts, and card totals for
        the user.
      produces:
      - application/json
      responses:
        "200":
          description: Retention stats
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get retention stats
      tags:
      - reviews
  /sessions/{id}/complete:
    post:
      description: Marks a quiz session as completed. If answers are provided, it
        triggers AI evaluation.
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Performance data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CompleteSessionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Completion confirmation
          schema:
            $ref: '#/definitions/internal_handlers.CompleteSessionResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete a session
      tags:
      - sessions
  /sessions/{id}/socratic:
    post:
      consumes:
      - application/json
      description: Evaluates a free-text answer conceptually and returns a "Why?"
        follow-up the user must answer before the explanation is revealed.
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      - description: The question and the user's answer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SocraticFollowUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: Socratic follow-up
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a Socratic follow-up on a text answer
      tags:
      - sessions
  /sessions/activity:
    get:
      description: Returns the last 7 days activity of the user.
      produces:
      - application/json
      responses:
        "200":
          description: Activity data
          schema:
            $ref: '#/definitions/internal_handlers.GetUserActivityResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get last 7 days activity
      tags:
      - sessions
  /sessions/start:
    post:
      consumes:
      - application/json
      description: Creates a new quiz session. If lesson_id is provided, it uses existing
        questions. Otherwise, it generates questions via AI.
      parameters:
      - description: Session configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.StartSessionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Session ID and questions
          schema:
            $ref: '#/definitions/internal_handlers.StartSessionResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a quiz session
      tags:
      - sessions
...
  /topics/{id}/review-cards/generate:
    post:
      description: Uses AI to generate spaced-repetition flashcards from the topic's
        roadmap and inserts them into the review queue.
      parameters:
      - description: Topic UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Number of generated cards
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Generation error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate review cards for a topic
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"